pyproject.toml
[tool.poetry]
name = "pyul"
version = "0.1.0"
description = ""
authors = ["elias <elias.jm.garcia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
antlr4-python3-runtime = "4.8"

# TODO: these belong to Eurus' python scripts, do we need them?
pysha3 = "1.0.2"
eth-abi = "2.1.1"

# Development dependencies
pytest = { version = "^7.1", optional = true }
flake8 = { version = "^4.0.1", optional = true }
mypy = { version = "0.961", optional = true } # poetry2nix fails on >= 0.971

[tool.poetry.extras]
dev = ["flake8", "mypy", "pytest"]

[tool.poetry.scripts]
pyul = "pyul.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.8"
exclude = ["src/utils"]

[[tool.mypy.overrides]]
module = "sha3"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "antlr4"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "utils.yul_parser"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "utils.yul_translator"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "utils.YulAntlr"
ignore_missing_imports = true